first working version (for sdoc search)
