+++ backend/src/api/endpoints/memo.py
@@ -5,7 +5,6 @@ from sqlalchemy.orm import Session
 
 from api.dependencies import get_current_user, get_db_session
 from api.validation import Validate
-from app.core.analysis.memo import MemoColumns, memo_info, memo_search
 from app.core.authorization.authz_user import AuthzUser
 from app.core.data.crud import Crud
 from app.core.data.crud.memo import crud_memo
@@ -19,9 +18,10 @@ from app.core.data.dto.memo import (
 )
 from app.core.data.dto.search import PaginatedElasticSearchDocumentHits
 from app.core.data.orm.util import get_parent_project_id
-from app.core.filters.columns import ColumnInfo
+from app.core.filters.column_info import ColumnInfo
 from app.core.filters.filtering import Filter
 from app.core.filters.sorting import Sort
+from app.core.search.memo_search.memo import MemoColumns, memo_info, memo_search
 
 router = APIRouter(
     prefix="/memo", dependencies=[Depends(get_current_user)], tags=["memo"]
