+++ backend/src/app/core/data/dto/search.py
@@ -1,14 +1,10 @@
 from datetime import datetime
-from typing import List, Optional, Union
+from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
 from app.core.data.dto.dto_base import UpdateDTOBase
 from app.core.data.dto.memo import AttachedObjectType
-from app.core.data.orm.source_document import SourceDocumentORM
-from app.core.filters.columns import AbstractColumns
-from app.core.filters.filtering import Filter
-from app.core.filters.filtering_operators import FilterOperator, FilterValueType
 
 
 class ElasticSearchDocumentCreate(BaseModel):
@@ -102,118 +98,6 @@ class PaginatedElasticSearchDocumentHits(BaseModel):
     )
     total_results: int = Field(
         description="The total number of hits. Used for pagination."
-    )
-
-
-class SearchColumns(str, AbstractColumns):
-    SOURCE_DOCUMENT_TYPE = "SC_SOURCE_DOCUMENT_TYPE"
-    SOURCE_DOCUMENT_FILENAME = "SC_SOURCE_DOCUMENT_FILENAME"
-    DOCUMENT_TAG_ID_LIST = "SC_DOCUMENT_TAG_ID_LIST"
-    CODE_ID_LIST = "SC_CODE_ID_LIST"
-    USER_ID_LIST = "SC_USER_ID_LIST"
-    SPAN_ANNOTATIONS = "SC_SPAN_ANNOTATIONS"
-
-    def get_filter_column(self, **kwargs):
-        subquery_dict = kwargs["subquery_dict"]
-
-        match self:
-            case SearchColumns.SOURCE_DOCUMENT_FILENAME:
-                return SourceDocumentORM.filename
-            case SearchColumns.SOURCE_DOCUMENT_TYPE:
-                return SourceDocumentORM.doctype
-            case SearchColumns.DOCUMENT_TAG_ID_LIST:
-                return subquery_dict[SearchColumns.DOCUMENT_TAG_ID_LIST]
-            case SearchColumns.CODE_ID_LIST:
-                return subquery_dict[SearchColumns.CODE_ID_LIST]
-            case SearchColumns.USER_ID_LIST:
-                return subquery_dict[SearchColumns.USER_ID_LIST]
-            case SearchColumns.SPAN_ANNOTATIONS:
-                return subquery_dict[SearchColumns.SPAN_ANNOTATIONS]
-
-    def get_filter_operator(self) -> FilterOperator:
-        match self:
-            case SearchColumns.SOURCE_DOCUMENT_FILENAME:
-                return FilterOperator.STRING
-            case SearchColumns.SOURCE_DOCUMENT_TYPE:
-                return FilterOperator.ID
-            case SearchColumns.DOCUMENT_TAG_ID_LIST:
-                return FilterOperator.ID_LIST
-            case SearchColumns.CODE_ID_LIST:
-                return FilterOperator.ID_LIST
-            case SearchColumns.USER_ID_LIST:
-                return FilterOperator.ID_LIST
-            case SearchColumns.SPAN_ANNOTATIONS:
-                return FilterOperator.ID_LIST
-
-    def get_filter_value_type(self) -> FilterValueType:
-        match self:
-            case SearchColumns.SOURCE_DOCUMENT_FILENAME:
-                return FilterValueType.INFER_FROM_OPERATOR
-            case SearchColumns.SOURCE_DOCUMENT_TYPE:
-                return FilterValueType.DOC_TYPE
-            case SearchColumns.DOCUMENT_TAG_ID_LIST:
-                return FilterValueType.TAG_ID
-            case SearchColumns.CODE_ID_LIST:
-                return FilterValueType.CODE_ID
-            case SearchColumns.USER_ID_LIST:
-                return FilterValueType.USER_ID
-            case SearchColumns.SPAN_ANNOTATIONS:
-                return FilterValueType.SPAN_ANNOTATION
-
-    def get_sort_column(self, **kwargs):
-        match self:
-            case SearchColumns.SOURCE_DOCUMENT_FILENAME:
-                return SourceDocumentORM.filename
-            case SearchColumns.SOURCE_DOCUMENT_TYPE:
-                return SourceDocumentORM.doctype
-            case SearchColumns.DOCUMENT_TAG_ID_LIST:
-                return None
-            case SearchColumns.CODE_ID_LIST:
-                return None
-            case SearchColumns.USER_ID_LIST:
-                return None
-            case SearchColumns.SPAN_ANNOTATIONS:
-                return None
-
-    def get_label(self) -> str:
-        match self:
-            case SearchColumns.SOURCE_DOCUMENT_FILENAME:
-                return "Document name"
-            case SearchColumns.SOURCE_DOCUMENT_TYPE:
-                return "Type"
-            case SearchColumns.DOCUMENT_TAG_ID_LIST:
-                return "Tags"
-            case SearchColumns.CODE_ID_LIST:
-                return "Code"
-            case SearchColumns.USER_ID_LIST:
-                return "Annotated by"
-            case SearchColumns.SPAN_ANNOTATIONS:
-                return "Span annotations"
-
-
-class SimSearchQuery(BaseModel):
-    proj_id: int = Field(
-        description="The ID of the Project the SourceDocuments have to belong to."
-    )
-    query: Union[str, List[str], int] = Field(
-        description=(
-            "The query term. This can be either a single string, "
-            "a list of strings for which the average embedding gets computed, "
-            "or an integer which is interpreted as the SDoc ID of an Image."
-        )
-    )
-    top_k: int = Field(
-        description="The number of results to return.",
-        default=100,
-    )
-    threshold: float = Field(
-        description="The minimum distance to use for the sim search.",
-        default=0.0,
-        ge=0.0,
-        le=1.0,
-    )
-    filter: Filter[SearchColumns] = Field(
-        description="Only return documents matching this filter"
     )
