+++ backend/src/app/core/db/sql_utils.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Integer, func
+from sqlalchemy.dialects.postgresql import ARRAY, array_agg
+from sqlalchemy.orm import InstrumentedAttribute
+
+
+def aggregate_ids(column: InstrumentedAttribute, label: str):
+    return func.array_remove(
+        array_agg(func.distinct(column), type_=ARRAY(Integer)),
+        None,
+        type_=ARRAY(Integer),
+    ).label(label)
