+++ backend/src/app/core/search/span_search/annotated_segments.py
@@ -1,9 +1,5 @@
 from typing import List, Optional
 
-from sqlalchemy import Integer, func
-from sqlalchemy.dialects.postgresql import ARRAY, array_agg
-from sqlalchemy.orm import InstrumentedAttribute
-
 from app.core.data.crud.project_metadata import crud_project_meta
 from app.core.data.doc_type import DocType
 from app.core.data.dto.analysis import (
@@ -12,6 +8,7 @@ from app.core.data.dto.analysis import (
 )
 from app.core.data.dto.code import CodeRead
 from app.core.data.dto.document_tag import DocumentTagRead
+from app.core.data.dto.project_metadata import ProjectMetadataRead
 from app.core.data.dto.source_document import SourceDocumentRead
 from app.core.data.orm.annotation_document import AnnotationDocumentORM
 from app.core.data.orm.code import CodeORM
@@ -23,7 +20,8 @@ from app.core.data.orm.span_annotation import SpanAnnotationORM
 from app.core.data.orm.span_text import SpanTextORM
 from app.core.data.orm.user import UserORM
 from app.core.db.sql_service import SQLService
-from app.core.filters.columns import (
+from app.core.db.sql_utils import aggregate_ids
+from app.core.filters.column_info import (
     AbstractColumns,
     ColumnInfo,
 )
@@ -31,14 +29,6 @@ from app.core.filters.filtering import Filter, apply_filtering
 from app.core.filters.filtering_operators import FilterOperator, FilterValueType
 from app.core.filters.pagination import apply_pagination
 from app.core.filters.sorting import Sort, apply_sorting
-
-
-def aggregate_ids(column: InstrumentedAttribute, label: str):
-    return func.array_remove(
-        array_agg(func.distinct(column), type_=ARRAY(Integer)),
-        None,
-        type_=ARRAY(Integer),
-    ).label(label)
 
 
 class AnnotatedSegmentsColumns(str, AbstractColumns):
@@ -129,9 +119,15 @@ def find_annotated_segments_info(
     project_id,
 ) -> List[ColumnInfo[AnnotatedSegmentsColumns]]:
     with SQLService().db_session() as db:
-        metadata_column_info = crud_project_meta.create_metadata_column_info(
-            db=db, project_id=project_id, allowed_doctypes=[DocType.text]
-        )
+        project_metadata = [
+            ProjectMetadataRead.model_validate(pm)
+            for pm in crud_project_meta.read_by_project(db=db, proj_id=project_id)
+        ]
+        metadata_column_info = [
+            ColumnInfo.from_project_metadata(pm)
+            for pm in project_metadata
+            if pm.doctype in [DocType.text]
+        ]
 
     return [
         ColumnInfo[AnnotatedSegmentsColumns].from_column(column)
@@ -220,9 +216,9 @@ def find_annotated_segments(
             )
         )
 
-        query = apply_filtering(query=query, filter=filter, db=db)
+        query = apply_filtering(query=query, filter=filter, subquery_dict={})
 
-        query = apply_sorting(query=query, sorts=sorts, db=db)
+        query = apply_sorting(query=query, sorts=sorts, subquery_dict={})
 
         if page is not None and page_size is not None:
             query = query.order_by(
